# EnterpriseWarehouse.Wasm/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EnterpriseWarehouse.API",
    "version": "1.0"
  },
  "paths": {
    "/api/Cell": {
      "get": {
        "tags": [
          "Cell"
        ],
        "summary": "Возвращает список всех ячеек склада.",
        "responses": {
          "200": {
            "description": "Список ячеек успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CellDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CellDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CellDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cell"
        ],
        "summary": "Добавляет новую ячейку на склад.",
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.CellCreateDTO, содержащий информацию о новой ячейке.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CellCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CellCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CellCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ячейка успешно добавлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка при добавлении ячейки."
          }
        }
      }
    },
    "/api/Cell/{id}": {
      "get": {
        "tags": [
          "Cell"
        ],
        "summary": "Получает информацию о ячейке склада по её идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор ячейки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ячейка найдена и информация успешно возвращена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ячейка с указанным идентификатором не найдена."
          }
        }
      },
      "put": {
        "tags": [
          "Cell"
        ],
        "summary": "Обновляет информацию о существующей ячейке.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор ячейки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.CellCreateDTO, содержащий обновлённые данные ячейки.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CellCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CellCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CellCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ячейка успешно обновлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ячейка с указанным идентификатором не найдена."
          }
        }
      },
      "delete": {
        "tags": [
          "Cell"
        ],
        "summary": "Удаляет ячейку склада по её идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор ячейки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ячейка успешно удалена."
          },
          "404": {
            "description": "Ячейка с указанным идентификатором не найдена."
          }
        }
      }
    },
    "/api/Organization": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Возвращает список всех организаций.",
        "responses": {
          "200": {
            "description": "Список организаций успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "summary": "Добавляет новую организацию.",
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.OrganizationCreateDTO, содержащий данные новой организации.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Организация успешно добавлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organization/{id}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Возвращает информацию об организации по её идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор организации.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Организация найдена и информация успешно возвращена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              }
            }
          },
          "404": {
            "description": "Организация с указанным идентификатором не найдена."
          }
        }
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "summary": "Обновляет существующую организацию.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор обновляемой организации.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.OrganizationCreateDTO, содержащий обновлённые данные организации.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Организация успешно обновлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDTO"
                }
              }
            }
          },
          "404": {
            "description": "Организация с указанным идентификатором не найдена."
          }
        }
      },
      "delete": {
        "tags": [
          "Organization"
        ],
        "summary": "Удаляет организацию по её идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор организации для удаления.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Организация успешно удалена."
          },
          "404": {
            "description": "Организация с указанным идентификатором не найдена."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CellCreateDTO": {
        "required": [
          "productId",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "Идентификатор товара",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "description": "Количество товара в ячейке",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания ячейки"
      },
      "CellDTO": {
        "required": [
          "id",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор ячейки",
            "format": "int32"
          },
          "product": {
            "$ref": "#/components/schemas/ProductDTO"
          },
          "quantity": {
            "type": "integer",
            "description": "Количество товара в ячейке",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO ячейки"
      },
...
      "OrganizationCreateDTO": {
        "required": [
          "address",
          "geometry",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "geometry": {
            "type": "string",
            "description": "Геометрия в виде полигона",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания организации"
      },
      "OrganizationDTO": {
        "required": [
          "address",
          "geometry",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "geometry": {
            "type": "string",
            "description": "Геометрия в виде полигона",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO организации"
      },
...
      "ProductDTO": {
        "required": [
          "code",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "code": {
            "type": "integer",
            "description": "Код товара",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для товара"
      },
...
    }
  }
}